Cleansing 'registered' and ARFCOM forced terminology
Arfcom mods forced usage of "RDIAS" because of the Shotgun News ATF entrapment ad that ran for 30 years in the back of the mag. We're making *unregistered sears* here. We're modeling it on a Drop-In Auto Sear, not a registered one. Registration defeats the purpose of FOSScad. Fuck ARFcom. Fuck registration.